--- 信息科学与工程学院/Python 程序设计与应用/README.md
@@ -1,48 +1,0 @@
-# 资料链接
-
-http://cs.e.ecust.edu.cn/download/fb0d9c0f79c2a92815a79589dc3332ce
-http://cs.e.ecust.edu.cn/download/81b105cef99e262f18c2393112b5848f
-http://cs.e.ecust.edu.cn/download/99444371d555929eb21a72e144d3994a
-http://cs.e.ecust.edu.cn/download/14b78871f76b304c6aaf7ff2cea69ac2
-http://cs.e.ecust.edu.cn/download/5722d639249cb448698f5977b89f2116
-http://cs.e.ecust.edu.cn/download/b48fd8d0a084b0b8d54ab49303ad90c6
-http://cs.e.ecust.edu.cn/download/1008345c0e30639156d83f9441dd25a0
-http://cs.e.ecust.edu.cn/download/a9cbea49124b9c59d4b306e1161dd432
-http://cs.e.ecust.edu.cn/download/77fe428f570c6d3de969f268f3276029
-http://cs.e.ecust.edu.cn/download/4940bdb186b23b889cdb2525ad94e4db
-
-http://cs.e.ecust.edu.cn/download/4906e7b74ec87d04e6f825297621817b
-http://cs.e.ecust.edu.cn/download/6545f75f27df36511bce5f10c8fc04df
-http://cs.e.ecust.edu.cn/download/7ddbf01793c4ddc4f81de74605fc6471
-http://cs.e.ecust.edu.cn/download/dd5476798b0f3d3a9c8c86bd741af14e
-http://cs.e.ecust.edu.cn/download/281336a7baa2826ae5e368e25399dda2
-http://cs.e.ecust.edu.cn/download/c6ba84a67240941a90ad3cc151f75f89
-http://cs.e.ecust.edu.cn/download/71b73b544779fb3dde7bbfad1f391a6f
-http://cs.e.ecust.edu.cn/download/f46a8adf4f29b750c85ea37cf18117e7
-http://cs.e.ecust.edu.cn/download/3ca0ac7c225eb52debee7944f20c04cb
-http://cs.e.ecust.edu.cn/download/4ba25359a6788aefe34fff7b2681462c
-
-http://cs.e.ecust.edu.cn/download/b25775fddccae53f766063a618566d4a
-http://cs.e.ecust.edu.cn/download/ad79f163911b934e41806cfe25a0d85f
-http://cs.e.ecust.edu.cn/download/845e95331507aa6067491da055dcdde4
-http://cs.e.ecust.edu.cn/download/5d954518982be3518128d7d8b450a2cb
-http://cs.e.ecust.edu.cn/download/c6f0c7db7a9bc50d546c34bad9e8d3e5
-http://cs.e.ecust.edu.cn/download/4bcd226e2e453ef4bea261f916963a71
-http://cs.e.ecust.edu.cn/download/4737409414d344c351230a8afbbdb809
-http://cs.e.ecust.edu.cn/download/171ca7d45c37626844c2a4bc7e3d1731
-http://cs.e.ecust.edu.cn/download/7c9bb98e7b5baad8fa3e130dd030db60
-http://cs.e.ecust.edu.cn/download/7941a549c0b9e77554d67ddbd00016db
-
-http://cs.e.ecust.edu.cn/download/018cbef4e2ff0cec75674e267bdb7612
-http://cs.e.ecust.edu.cn/download/f3685c9838c8647a17bdd8bde4896bef
-http://cs.e.ecust.edu.cn/download/e32882580248bd5622f947e6dfe62472
-http://cs.e.ecust.edu.cn/download/97b692e39bbdac09c9ef257987d8ce3b
-http://cs.e.ecust.edu.cn/download/d948f01dc755b829fbd475cf4e7ba7d6
-http://cs.e.ecust.edu.cn/download/1b8d1016ad8ccc1f71999855074349c2
-http://cs.e.ecust.edu.cn/download/32b7935a60a189b489ec955ff0512ce9
-http://cs.e.ecust.edu.cn/download/ec2c0a507fcd862620179c00a1ac2eca
-http://cs.e.ecust.edu.cn/download/a682bbc9b1b2fe105df2b6fdb8f104a3
-http://cs.e.ecust.edu.cn/download/70443a089f8177f1439015d46fd60927
-
-http://cs.e.ecust.edu.cn/download/4e655a9001e2498406ca08c94ecd34d6
-http://cs.e.ecust.edu.cn/download/b8af8e84154825ed1adef129d53b7c58
