--- 信息科学与工程学院/Python程序设计-高建良/ex1_print.py
@@ -1,12 +1,0 @@
-print(
-    '''
-  *****
-  *   *
-  *****
-
-  *****
-*********
-   * *
-**     **
-'''
-)
