+++ 信息科学与工程学院/python-程序设计/Python程序设计-高建良/ex1-temperature.py
@@ -1,0 +1,2 @@
+t_f = float(input("请输入华氏温度："))
+print("对应的摄氏温度为：", (t_f-32)/1.8)
