+++ 信息科学与工程学院/python-程序设计/Python程序设计-高建良/ex1_print.py
@@ -1,0 +1,12 @@
+print(
+    '''
+  *****
+  *   *
+  *****
+
+  *****
+*********
+   * *
+**     **
+'''
+)
