+++ 信息科学与工程学院/python-程序设计/Python程序设计-高建良/ex4_scores.txt
@@ -1,0 +1,20 @@
+66
+45
+18
+40
+49
+23
+56
+29
+94
+63
+36
+37
+32
+35
+74
+92
+98
+92
+63
+8
